[doc] replace usages of `./bazelisk.sh` with `bazel` in Getting Started
In Getting Started Section 2.4 "Building (and Testing) Software", the
bazel installation section recommends either using the wrapper
`./bazelisk.sh` in place of `bazel`, or aliasing `bazel` to execute this
wrapper script.

This changes all the usages of `./bazelisk.sh` in the Getting Started
docs to `bazel`, to be consistent for those who have installed bazel
manually and those who use the script aliased.

Signed-off-by: Alice Ziuziakowska <a.ziuziakowska@lowrisc.org>